Move structure to PSR-4 and update README and libs
