+++ README.md
@@ -5,11 +5,25 @@
 
 #### Dependencies:
 - PHP >= 5.6
-- [FastRoute](https://github.com/nikic/FastRoute)
 
 ### Installation
 - `composer install`
+- Run development server with `php -S localhost:8000 -t .`
+- Run examples in browser:
+```
+// Web APP 
+http://localhost:8000/index.html
 
+// API: Generate Roman numeral 
+http://localhost:8000/api/generate/123
+
+// API: Parse Roman numeral 
+http://localhost:8000/api/parse/CXXIII
+
+```
+### Tests
+Run
+`vendor/bin/phpunit -c phpunit.xml tests`
 ### Live demo
 
 <http://roman.alexlv.com>
