+++ composer.json
@@ -1,18 +1,19 @@
 {
   "require": {
-    "nikic/fast-route": "*"
+    "nikic/fast-route": "*",
+    "php": ">=5.6.0"
   },
   "require-dev": {
     "symfony/var-dumper": "*",
-    "phpunit/phpunit": "5.0.*",
+    "phpunit/phpunit": "^6.1",
     "squizlabs/php_codesniffer": "2.*"
   },
   "autoload": {
     "classmap": [
     ],
     "psr-4": {
-      "": "",
-      "RomanNumerals\\Testing\\": "tests"
+      "RomanNumerals\\": "app",
+      "RomanNumerals\\Tests\\": "tests"
     }
   }
 }
