+++ index.php
@@ -1,6 +1,11 @@
 <?php
 
-namespace RomanNumerals;
+use RomanNumerals\Container;
+use RomanNumerals\Controllers\Controller;
+use RomanNumerals\Controllers\ConvertController;
+use RomanNumerals\Libs\RomanNumeralGenerator;
+use RomanNumerals\Router;
+use RomanNumerals\Application;
 
 error_reporting(E_ALL);
 ini_set('display_errors', 1);
@@ -11,24 +16,24 @@ $routes = [
     [
         'GET',
         '/api/generate/{arabicNumber:\w+}',
-        'RomanNumerals\ConvertController@generate'
+        ConvertController::class.'@generate'
     ],
     [
         'GET',
         '/api/parse/{romanNumber:\w+}',
-        'RomanNumerals\ConvertController@parse'
+        ConvertController::class.'@parse'
     ],
 ];
 
 $container = new Container();
 
-$container->bind('RomanNumerals\ConvertController', function () {
+$container->bind(ConvertController::class, function () {
     $object = new ConvertController();
     $object->setRomanNumeralGenerator(new RomanNumeralGenerator());
     return $object;
 });
 
-$container->bind('RomanNumerals\Controller', function () {
+$container->bind(Controller::class, function () {
     return new ConvertController();
 });
